requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
blas=2.131=openblas
blas-devel=3.9.0=31_h1ea3ea9_openblas
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.8.3=hbd8a1cb_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
expat=2.7.1=h6a678d5_0
faiss-gpu=1.7.2=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.59.2=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.3.1=pypi_0
h5py=3.8.0=pypi_0
hf-xet=1.1.9=pypi_0
huggingface-hub=0.34.4=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.5.2=pypi_0
kiwisolver=1.4.9=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libblas=3.9.0=31_h59b9bed_openblas
libcblas=3.9.0=31_he106b2a_openblas
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=31_h7ac8fdf_openblas
liblapacke=3.9.0=31_he2f377e_openblas
libopenblas=0.3.29=ha39b09d_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
lightning-utilities=0.15.2=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.6=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.6.4=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.3=pypi_0
nomkl=3.0=0
numpy=1.24.3=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.20.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.17.2=pypi_0
openblas=0.3.29=h06a4308_0
openblas-devel=0.3.29=ha39b09d_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=pypi_0
pip=25.2=pyhc872135_0
propcache=0.3.2=pypi_0
protobuf=3.20.3=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pygments=2.19.2=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.0=pypi_0
pytorch-lightning=2.5.5=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.3=hc2a1206_0
regex=2025.9.1=pypi_0
requests=2.32.5=pypi_0
rich=13.3.5=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.15.3=py310h5e4e545_0
sentencepiece=0.2.1=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.3=pypi_0
threadpoolctl=3.6.0=pypi_0
timm=1.0.19=pypi_0
tk=8.6.15=h54e0aa7_0
torch=2.4.0+cu118=pypi_0
torchaudio=2.4.0+cu118=pypi_0
torchmetrics=1.8.2=pypi_0
torchvision=0.19.0+cu118=pypi_0
tqdm=4.67.1=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.45.1=py310h06a4308_0
xformers=0.0.27.post2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
